refactor(staking): 𝚨 `ValidatorUptimeTracker::track_uptime()`
