+++ crates/core/component/stake/src/component/validator_handler.rs
@@ -5,3 +5,6 @@ pub mod validator_store;
 pub use validator_store::ValidatorDataRead;
 pub(crate) use validator_store::ValidatorDataWrite;
 pub(crate) use validator_store::ValidatorPoolTracker;
+
+pub mod uptime_tracker;
+pub use uptime_tracker::ValidatorUptimeTracker;
